Rename to RatingBadge and update border radius value
