+++ src/index.js
@@ -7,13 +7,13 @@ export { default as Divider } from './Divider'
 export { default as GreenButton } from './GreenButton'
 export { default as RedButton } from './RedButton'
 export { default as OutlineButton } from './OutlineButton'
-export { default as GuestRating } from './GuestRating'
 export { default as Flex } from './Flex'
 export { default as Heading } from './Heading'
 export { default as Hide } from './Hide'
 export { default as Icon } from './Icon'
 export { default as Image } from './Image'
 export { default as Link } from './Link'
+export { default as RatingBadge } from './RatingBadge'
 export { default as Text } from './Text'
 export { default as Truncate } from './Truncate'
 export { default as Container } from './Container'
