+++ src/theme.js
@@ -114,7 +114,7 @@ export const radius = '2px'
 
 // todo: wes can't decide on a size
 // styled-system's `borderRadius` function can hook into the `radii` object/array
-export const radii = [0, 2, 4, 6]
+export const radii = [0, 2, 6]
 
 export const maxContainerWidth = '1280px'
